jpms: Refactor handlebars.helpers
