+++ handlebars-helpers/src/main/java/com/github/jknack/handlebars/helpers/JodaHelper.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.jknack.handlebars.helper;
+package com.github.jknack.handlebars.helpers;
 
 import java.io.IOException;
